route /getFromCart working and empty cart on logout

+++ angular-cli/src/app/cart/cart-service.service.ts
@@ -1,27 +1,26 @@
 import { Injectable } from '@angular/core';
-import { Product} from "../product/product.model";
-import {Subject} from "rxjs/Subject";
-import { ProductDatabaseService } from "../product/services/product-database.service";
-import { AuthService } from "../Auth/auth.service";
+import { Product} from '../product/product.model';
+import {Subject} from 'rxjs/Subject';
+import { ProductDatabaseService } from '../product/services/product-database.service';
+import { AuthService } from '../Auth/auth.service';
 @Injectable()
 export class CartService {
 
- constructor(private productDatabaseService:ProductDatabaseService,private authService:AuthService){}
- private cartProduct:Product[]=[];
- onProductsChanged = new Subject<Product[]>();
+  private cartProduct: Product[]= [];
+  onProductsChanged = new Subject<Product[]>();
+ constructor(private productDatabaseService: ProductDatabaseService, private authService: AuthService) {}
 
-  fillCart(data: Product) {
+  fillCart(data: Product, quantity) {
     this.cartProduct.push(data);
 //     const token = localStorage.getItem("token");
 // console.log(token);
     if (!this.authService.isLoggedIn()) {
       localStorage.setItem('Cart', JSON.stringify(this.cartProduct));
-    }
-    else {
-      console.log("token is not null");
-      this.productDatabaseService.addCartData(data._id).subscribe(
-        () =>{
-          console.log("added to database");
+    }else {
+      console.log('token is not null');
+      this.productDatabaseService.addCartData(data._id,quantity).subscribe(
+        () => {
+          console.log('added to database');
         }
       )
     }
@@ -31,28 +30,35 @@ export class CartService {
   }
 
 
-  fillCartArray(data:Product[]){
+  fillCartArray(data: Product[]) {
+    console.log(data);
     this.cartProduct = data;
-    localStorage.setItem('Cart', JSON.stringify(this.cartProduct));
+    // localStorage.setItem('Cart', JSON.stringify(this.cartProduct));
     this.onProductsChanged.next(this.cartProduct.slice());
 
   }
 
-  getCart(){
+  getCart() {
     return this.cartProduct.slice();
   }
 
-  removeProduct(index:number)
-  {
-    const removedProduct=this.cartProduct.splice(index,1);
+  removeProduct(index: number) {
+    const removedProduct = this.cartProduct.splice(index, 1);
 
     // localStorage.setItem('Cart', JSON.stringify(this.cartProduct));
-    this.productDatabaseService.removeCartData(removedProduct[0]._id).subscribe(
-      (data) => {
-        console.log(data);
-      }
-    )
+    if (this.authService.isLoggedIn) {
+      this.productDatabaseService.removeCartData(removedProduct[0]._id).subscribe(
+        (data) => {
+          console.log(data);
+        }
+      );
+    }
 
+    this.onProductsChanged.next(this.cartProduct.slice());
+  }
+
+  emptyCart() {
+    this.cartProduct = [];
     this.onProductsChanged.next(this.cartProduct.slice());
   }
